product service tests

+++ src/main/java/com/dscommerce/dscommerce/services/ProductService.java
@@ -38,15 +38,15 @@ public class ProductService {
     @Transactional(readOnly = true)
     public Page<ProductMinDTO> findAll(String name, Pageable pageable) {
         Page<Product> result = repository.searchByName(name, pageable);
-        return result.map(x -> new ProductMinDTO(x));
+        return result.map(ProductMinDTO::new);
     }
 
     @Transactional
     public ProductDTO insert(ProductDTO input) {
         Product entity = new Product();
         copyDtoToEntity(input, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        Product data = repository.save(entity);
+        return new ProductDTO(data);
     }
 
     @Transactional
@@ -63,10 +63,11 @@ public class ProductService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
         try {
             repository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new ResourceNotFoundException("Recurso não encontrado");
         } catch (DataIntegrityViolationException e) {
             throw new DatabaseException("Falha de integridade referencial");
         }
